Amqp component based on amqp interop.
